chore(builder): handle shared data on init
